pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ollama_agent"
version = "0.1.0"
description = "Configurável OllamaAgent em Python para consumo rápido do Ollama"
authors = [ { name = "OllamaAgent", email = "dev@example.com" } ]
dependencies = [
    "requests>=2.28.0",
    "pyzabbix>=1.3.0",
    "tenacity>=8.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "mcp>=1.0.0",
    "pysnmp>=4.4.12"
]

[tool.setuptools.packages.find]
where = ["."]
